# docs/swagger.yaml
definitions:
...
  dto.AddressDto:
    properties:
      city:
        type: string
      id:
        type: string
      state:
        type: string
      street:
        type: string
      zip_code:
        type: string
    type: object
...
  dto.CommunityDto:
    properties:
      address:
        $ref: '#/definitions/dto.AddressDto'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/dto.UserDtoOut'
    type: object
  dto.EventDto:
    properties:
      address:
        $ref: '#/definitions/dto.AddressDto'
      category:
        type: string
      community:
        $ref: '#/definitions/dto.CommunityDto'
      description:
        type: string
      duration_min:
        type: integer
      id:
        type: string
      max_slots:
        type: integer
      meeting_link:
        type: string
      owner:
        $ref: '#/definitions/dto.UserDtoOut'
      start_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  dto.PageableCommunityDto:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CommunityDto'
        type: array
      has_next:
        type: boolean
    type: object
  dto.PageableEventDto:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.EventDto'
        type: array
      has_next:
        type: boolean
    type: object
...
  dto.RegisterCommunityDto:
    properties:
      address_id:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
  dto.RegisterEventDto:
    properties:
      address_id:
        type: string
      category:
        type: string
      community_id:
        type: string
      description:
        type: string
      duration_min:
        type: integer
      id:
        type: string
      max_slots:
        type: integer
      meeting_link:
        type: string
      owner_id:
        type: string
      start_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  dto.UserDtoOut:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
...
info:
  contact: {}
  title: AjudaDev Backend API
  version: "1.0"
paths:
  /v1/address/register:
    post:
      consumes:
      - application/json
      description: Cria um novo endereço no sistema
      parameters:
      - description: Dados do endereço
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/dto.AddressDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AddressDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Registra um novo endereço
      tags:
      - addresses
  /v1/community:
    get:
      consumes:
      - application/json
      description: Retorna todas as comunidades, com paginação e filtro por endereço
      parameters:
      - description: Página
        in: query
        name: page
        type: integer
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: ID do endereço
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PageableCommunityDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lista comunidades
      tags:
      - communities
  /v1/community/register:
    post:
      consumes:
      - application/json
      description: Cria uma nova comunidade no sistema
      parameters:
      - description: Dados da comunidade
        in: body
        name: community
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterCommunityDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterCommunityDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Registra uma nova comunidade
      tags:
      - communities
  /v1/event:
    get:
      consumes:
      - application/json
      description: Retorna todos os eventos, com paginação e filtros por comunidade,
        categoria, tipo, endereço, cidade e apenas futuros
      parameters:
      - description: Página
        in: query
        name: page
        type: integer
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: ID da comunidade
        in: query
        name: community_id
        type: string
      - description: Categoria (COMMUNITY_EVENT, MENTORING, WEBINAR)
        in: query
        name: category
        type: string
      - description: Formato (ONLINE, INPERSON, HYBRID)
        in: query
        name: type
        type: string
      - description: ID do endereço
        in: query
        name: address_id
        type: string
      - description: Cidade do endereço do evento
        in: query
        name: city
        type: string
      - description: Somente eventos futuros
        in: query
        name: upcoming
        type: boolean
      - description: 'ID do usuário (agenda: eventos com participação)'
        in: query
        name: user_id
        type: string
      - description: Papel da participação (com user_id)
        in: query
        name: role
        type: string
      - description: Status da participação (com user_id)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PageableEventDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lista eventos
      tags:
      - events
...
  /v1/event/{id}:
    delete:
      description: Arquiva o evento marcando deleted_at
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove evento (soft delete)
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Retorna o detalhe do evento com owner, endereço e comunidade (preloads)
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventDto'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Busca evento por id
      tags:
      - events
  /v1/event/register:
    post:
      consumes:
      - application/json
      description: Cria um novo evento no sistema. Para eventos INPERSON/HYBRID, address_id
        é obrigatório; para ONLINE, address_id deve ser nulo.
      parameters:
      - description: Dados do evento
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterEventDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterEventDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Registra um novo evento
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"